+++ relm4/src/component/shutdown_on_drop.rs
@@ -3,29 +3,29 @@ use tokio::sync::oneshot;
 /// A type that destroys an [`AsyncComponent`](crate::async_component::AsyncComponent)
 /// as soon as it is dropped.
 #[derive(Debug)]
-pub(crate) struct DestroyOnDrop {
+pub(super) struct ShutdownOnDrop {
     /// Sender used to indicate that the async component should shut down.
-    destroy_sender: Option<oneshot::Sender<()>>,
+    shutdown_event_sender: Option<oneshot::Sender<()>>,
 }
 
-impl DestroyOnDrop {
+impl ShutdownOnDrop {
     /// Creates a new [`DestroyOnDrop`] type.
     ///
     /// When this type is dropped, a message will be sent through the channel.
-    pub(crate) fn new(destroy_sender: oneshot::Sender<()>) -> Self {
+    pub(crate) fn new(shutdown_event_sender: oneshot::Sender<()>) -> Self {
         Self {
-            destroy_sender: Some(destroy_sender),
+            shutdown_event_sender: Some(shutdown_event_sender),
         }
     }
 
-    pub(super) fn deactivate(&mut self) {
-        self.destroy_sender = None;
+    pub(crate) fn deactivate(&mut self) {
+        self.shutdown_event_sender = None;
     }
 }
 
-impl Drop for DestroyOnDrop {
+impl Drop for ShutdownOnDrop {
     fn drop(&mut self) {
-        if let Some(sender) = self.destroy_sender.take() {
+        if let Some(sender) = self.shutdown_event_sender.take() {
             sender.send(()).ok();
         }
     }